+++ docs-markdown/changelog.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- Added Unit Test Coverage: code-controller.ts
 - Added Moniker to quick-pick-menu-controller.ts
 - Added Unit Test Coverage: include-controller.ts